requirements.txt
torch>=2.0.0
transformers>=4.35.0
hypothesis>=6.0.0
pytest>=7.0.0
pyyaml>=6.0
